backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.3
tzdata==2024.2
idna==3.10
gunicorn==23.0.0
django-environ